backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
langchain>=0.1.0
langgraph>=0.0.26
langchain-google-genai>=0.0.9
PyMuPDF>=1.23.0
python-docx>=1.1.0
python-dotenv>=1.0.0
chromadb>=0.4.15
langchain-community>=0.0.10
langchain-ollama>=0.1.0
pandas>=2.0.0
tqdm>=4.65.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
langchain-postgres>=0.0.9
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client   